pax-collector/pax-exporter/pyproject.toml
[tool.poetry]
name = "pax-exporter"
version = "0.1.0"
description = ""
authors = ["Sarah Gäumann <sarah.gaeumann@edu.hefr.ch>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
paho-mqtt = "^1.6.1"
prometheus-client = "^0.17.0"
pylint = "^2.17.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
